Sender behov for arbeidsgiverinformasjon, featuretogglet av i prod + dev

+++ src/main/kotlin/no/nav/helse/modell/VedtakDao.kt
@@ -19,7 +19,7 @@ internal class VedtakDao(private val dataSource: DataSource) {
         fom: LocalDate,
         tom: LocalDate,
         personRef: Int,
-        arbeidsgiverRef: Int,
+        arbeidsgiverRef: Long,
         speilSnapshotRef: Int
     ) = using(sessionOf(dataSource)) {
         @Language("PostgreSQL")
